Rename form validator and type
