+++ packages/app/src/features/dialogs/savings/deposit/logic/createTxOverview.ts
@@ -2,13 +2,13 @@ import { SavingsInfo } from '@/domain/savings-info/types'
 import { NormalizedUnitNumber } from '@/domain/types/NumericValues'
 import { Token } from '@/domain/types/Token'
 import { TokensInfo } from '@/domain/wallet/useTokens/TokenInfo'
-import { TokenWithBalanceFormNormalizedData } from '@/features/dialogs/common/logic/asset-balance/form'
+import { TransferAmountFormNormalizedData } from '@/features/dialogs/common/logic/transfer-amount/form'
 import { TxOverviewRouteItem } from '@/features/dialogs/common/types'
 import { raise } from '@/utils/assert'
 import { SavingsDialogTxOverview } from '../../common/types'
 
 export interface CreateTxOverviewParams {
-  formValues: TokenWithBalanceFormNormalizedData
+  formValues: TransferAmountFormNormalizedData
   tokensInfo: TokensInfo
   savingsInfo: SavingsInfo
   type: 'sdai' | 'susds'
@@ -50,7 +50,7 @@ export function createTxOverview({
 }
 
 export interface GetDepositRouteParams {
-  formValues: TokenWithBalanceFormNormalizedData
+  formValues: TransferAmountFormNormalizedData
   tokensInfo: TokensInfo
   savingsInfo: SavingsInfo
   savingsToken: Token
