+++ packages/app/src/features/dialogs/savings/migrate/common/logic/createTxOverview.ts
@@ -1,11 +1,11 @@
 import { SavingsInfo } from '@/domain/savings-info/types'
 import { Token } from '@/domain/types/Token'
 import { TokensInfo } from '@/domain/wallet/useTokens/TokenInfo'
-import { TokenWithBalanceFormNormalizedData } from '@/features/dialogs/common/logic/asset-balance/form'
+import { TransferAmountFormNormalizedData } from '@/features/dialogs/common/logic/transfer-amount/form'
 import { MigrateDialogTxOverview } from '../types'
 
 export interface CreateTxOverviewParams {
-  formValues: TokenWithBalanceFormNormalizedData
+  formValues: TransferAmountFormNormalizedData
   tokensInfo: TokensInfo
   outputToken: Token
   savingsDaiInfo: SavingsInfo
