+++ packages/app/src/features/dialogs/savings/migrate/common/logic/useMigrateDialog.ts
@@ -7,8 +7,8 @@ import { Token } from '@/domain/types/Token'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
 import { useTokensInfo } from '@/domain/wallet/useTokens/useTokensInfo'
 import { InjectedActionsContext, Objective } from '@/features/actions/logic/types'
-import { useDebouncedFormValues } from '@/features/dialogs/common/logic/asset-balance/form'
 import { AssetInputSchema, getFormFieldsForAssetBalanceDialog } from '@/features/dialogs/common/logic/form'
+import { useDebouncedFormValues } from '@/features/dialogs/common/logic/transfer-amount/form'
 import { FormFieldsForDialog, PageState, PageStatus } from '@/features/dialogs/common/types'
 import { assert } from '@/utils/assert'
 import { zodResolver } from '@hookform/resolvers/zod'
