+++ packages/app/src/features/dialogs/savings/withdraw/logic/createObjectives.ts
@@ -1,11 +1,11 @@
 import { TokenWithBalance } from '@/domain/common/types'
 import { CheckedAddress } from '@/domain/types/CheckedAddress'
 import { WithdrawFromSavingsObjective } from '@/features/actions/flavours/withdraw-from-savings/types'
-import { TokenWithBalanceFormNormalizedData } from '@/features/dialogs/common/logic/asset-balance/form'
+import { TransferAmountFormNormalizedData } from '@/features/dialogs/common/logic/transfer-amount/form'
 import { Mode } from '../types'
 
 export interface CreateObjectivesParams {
-  formValues: TokenWithBalanceFormNormalizedData
+  formValues: TransferAmountFormNormalizedData
   receiver: CheckedAddress | undefined
   mode: Mode
   savingsTokenWithBalance: TokenWithBalance
