+++ packages/app/src/features/dialogs/savings/withdraw/logic/useSavingsWithdrawDialog.ts
@@ -6,8 +6,8 @@ import { useSavingsTokens } from '@/domain/savings/useSavingsTokens'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
 import { TokensInfo } from '@/domain/wallet/useTokens/TokenInfo'
 import { InjectedActionsContext, Objective } from '@/features/actions/logic/types'
-import { useDebouncedFormValues } from '@/features/dialogs/common/logic/asset-balance/form'
 import { AssetInputSchema } from '@/features/dialogs/common/logic/form'
+import { useDebouncedFormValues } from '@/features/dialogs/common/logic/transfer-amount/form'
 import { FormFieldsForDialog, PageState, PageStatus } from '@/features/dialogs/common/types'
 import { raise } from '@/utils/assert'
 import { zodResolver } from '@hookform/resolvers/zod'
