+++ packages/app/src/features/farm-details/dialogs/stake/logic/validation.ts
@@ -1,6 +1,6 @@
-import { BalanceValidationIssue } from '@/features/dialogs/common/logic/asset-balance/validation'
+import { TransferAmountValidationIssue } from '@/features/dialogs/common/logic/transfer-amount/validation'
 
-export const validationIssueToMessage: Record<BalanceValidationIssue, string> = {
+export const validationIssueToMessage: Record<TransferAmountValidationIssue, string> = {
   'value-not-positive': 'Stake value should be positive',
   'exceeds-balance': 'Exceeds your balance',
 }
